tools/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1
psycopg2-binary==2.9.9  # For future DB interactions
